+++ HW-2-4-3/CMakeLists.txt
@@ -3,5 +3,7 @@ project(Ibragimov_CPP_2022)
 
 set(CMAKE_CXX_STANDARD 14)
 
+set(SOURCES LinkedListQueue.h LinkedListQueue.cpp)
+
 add_executable(Ibragimov_CPP_2022
         HW-2-4-3.cpp)
